requirements.txt
streamlit==1.41.1
streamlit-ace==0.1.1
pandas==2.2.3
rdflib==7.1.3
SPARQLWrapper==2.0.0
openai==0.28.0
google.generativeai==0.8.3